App: Add track to top of searchResults when removing from playlistTracks
Previous implementation only inserted into correct location when it was the last item on the playlist being removed and on the first search. Attempting to get the reinsertion working as expected for songs being removed in any order and across searches would have been very complex for little gain. Instead, simply add the removed song to the top of the current search results.