App: Create temporary access button
Add button at the top of the page. Pressing will scroll down to the bottom of the page where temporary user details will be displayed. The reason for this is to make it clear at first look that there are access details available. Without the button (or similar), the user might not scroll down far enough to get this information.