App Bug: Saving users first playlist does not update list
When a user clicks on GET PLAYLISTS they are presented with the 'no playlists' message. But then if they create a new playlist, the PlaylistList did not update. This is resolved by checking that the button was pressed rather than for the existence of playlists (as there was none but an update was still required).